feat: add get listing call

+++ src/main/java/org/example/auction_platform/repository/bid/BidRepository.java
@@ -3,5 +3,5 @@ package org.example.auction_platform.repository.bid;
 import org.example.auction_platform.repository.bid.entity.Bid;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BidRepository extends JpaRepository<Bid,Integer> {
+public interface BidRepository extends JpaRepository<Bid,Long> {
 }
