+++ src/main/java/org/example/auction_platform/repository/listing/FinishedListingRepository.java
@@ -3,5 +3,5 @@ package org.example.auction_platform.repository.listing;
 import org.example.auction_platform.repository.listing.entity.FinishedListing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FinishedListingRepository extends JpaRepository<FinishedListing,Integer> {
+public interface FinishedListingRepository extends JpaRepository<FinishedListing,Long> {
 }
