+++ src/main/java/org/example/auction_platform/repository/listing/OngoingListingRepository.java
@@ -3,5 +3,5 @@ package org.example.auction_platform.repository.listing;
 import org.example.auction_platform.repository.listing.entity.OngoingListing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OngoingListingRepository extends JpaRepository<OngoingListing,Integer> {
+public interface OngoingListingRepository extends JpaRepository<OngoingListing,Long> {
 }
